feat(sauth): perfect sauth impl , integrate the interceptor into base module

+++ ej-modules/ej-sauth/src/main/java/easy4j/module/sauth/domain/SecuritySession.java
@@ -108,7 +108,7 @@ public class SecuritySession {
     /**
      * 过期时间（秒为单位）
      */
-    private int expireTimeSeconds;
+    private long expireTimeSeconds;
 
 
     /**
